feat(order): payment processing iyzico

+++ Controllers/OrderController.cs
@@ -1,5 +1,8 @@
 using dotnet_store.Models;
 using dotnet_store.Services;
+using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,11 +15,17 @@ public class OrderController : Controller
 {
     private readonly ICartService _cartService;
     private readonly DataContext _context;
+    private readonly IConfiguration _configuration;
 
-    public OrderController(ICartService cartService, DataContext context)
+    public OrderController(
+        ICartService cartService,
+        DataContext context,
+        IConfiguration configuration
+    )
     {
         _cartService = cartService;
         _context = context;
+        _configuration = configuration;
     }
 
     public async Task<ActionResult> Index()
@@ -87,7 +96,7 @@ public class OrderController : Controller
             SiparisTarihi = DateTime.Now,
             ToplamFiyat = cart.Toplam(),
             OrderItems = cart
-                .CartItems.Select(i => new OrderItem
+                .CartItems.Select(i => new Models.OrderItem
                 {
                     Urun = i.Urun,
                     UrunId = i.UrunId,
@@ -97,11 +106,21 @@ public class OrderController : Controller
                 .ToList(),
         };
 
-        await _context.Orders.AddAsync(order);
-        _context.Carts.Remove(cart);
-        await _context.SaveChangesAsync();
+        var payment = await ProcessPayment(model, cart);
 
-        return RedirectToAction("Completed", new { orderId = order.Id });
+        if (payment.Status == "success")
+        {
+            await _context.Orders.AddAsync(order);
+            _context.Carts.Remove(cart);
+
+            await _context.SaveChangesAsync();
+
+            return RedirectToAction("Completed", new { orderId = order.Id });
+        }
+
+        ModelState.AddModelError("", payment.ErrorMessage);
+        ViewBag.Cart = cart;
+        return View(model);
     }
 
     public async Task<ActionResult> Completed(int orderId)
@@ -194,5 +213,84 @@ public class OrderController : Controller
             .ToList();
 
         return View(model);
+    }
+
+    private async Task<Payment> ProcessPayment(OrderCreateModel model, Cart cart)
+    {
+        Options options = new Options();
+        options.ApiKey = _configuration["PaymentAPI:APIKey"];
+        options.SecretKey = _configuration["PaymentAPI:SecretKey"];
+        options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+        CreatePaymentRequest request = new CreatePaymentRequest();
+        request.Locale = Locale.TR.ToString();
+        request.ConversationId = Guid.NewGuid().ToString();
+        request.Price = cart.AraToplam().ToString();
+        request.PaidPrice = cart.AraToplam().ToString();
+        request.Currency = Currency.TRY.ToString();
+        request.Installment = 1;
+        request.BasketId = "B67832";
+        request.PaymentChannel = PaymentChannel.WEB.ToString();
+        request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+        PaymentCard paymentCard = new PaymentCard();
+        paymentCard.CardHolderName = model.CartName;
+        paymentCard.CardNumber = model.CartNumber;
+        paymentCard.ExpireMonth = model.CartExpirationMonth;
+        paymentCard.ExpireYear = model.CartExpirationYear;
+        paymentCard.Cvc = model.CartCVV;
+        paymentCard.RegisterCard = 0;
+        request.PaymentCard = paymentCard;
+
+        Buyer buyer = new Buyer();
+        buyer.Id = "BY789";
+        buyer.Name = model.AdSoyad;
+        buyer.Surname = "Doe";
+        buyer.GsmNumber = model.Telefon;
+        buyer.Email = "email@email.com";
+        buyer.IdentityNumber = "74300864791";
+        buyer.LastLoginDate = "2015-10-05 12:43:35";
+        buyer.RegistrationDate = "2013-04-21 15:12:09";
+        buyer.RegistrationAddress = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
+        buyer.Ip = "85.34.78.112";
+        buyer.City = model.Sehir;
+        buyer.Country = "Turkey";
+        buyer.ZipCode = model.PostaKodu;
+        request.Buyer = buyer;
+
+        Address shippingAddress = new Address();
+        shippingAddress.ContactName = model.AdSoyad;
+        shippingAddress.City = model.Sehir;
+        shippingAddress.Country = "Turkey";
+        shippingAddress.Description = model.AdresSatiri;
+        shippingAddress.ZipCode = model.PostaKodu;
+        request.ShippingAddress = shippingAddress;
+
+        Address billingAddress = new Address();
+        billingAddress.ContactName = model.AdSoyad;
+        billingAddress.City = model.Sehir;
+        billingAddress.Country = "Turkey";
+        billingAddress.Description = model.AdresSatiri;
+        billingAddress.ZipCode = model.PostaKodu;
+        request.BillingAddress = billingAddress;
+
+        List<BasketItem> basketItems = new List<BasketItem>();
+
+        foreach (var item in cart.CartItems)
+        {
+            BasketItem basketItem = new BasketItem();
+            basketItem.Id = item.CartItemId.ToString();
+            basketItem.Name = item.Urun.UrunAdi;
+            basketItem.Category1 = "Phone";
+            basketItem.Category2 = "Accessories";
+            basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+            basketItem.Price = item.Urun.Fiyat.ToString();
+
+            basketItems.Add(basketItem);
+        }
+
+        request.BasketItems = basketItems;
+
+        return await Payment.Create(request, options);
     }
 }
